+++ CHANGELOG.md
@@ -4,12 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2019-09-30
+### Changed
+- Fixed mock auth errors
+
 ## [0.1.3] - 2019-09-29
 ### Added
 - ULA Library
 - ULA UI Library
 - ULA Ledger Library
 - Mock auth provider
+### Changed
 - Basic UI code just to show the button and modal
 - Updated Vue version